fix: moved 2 endpoints from admin and post controller to user controller.
