+++ web/src/main/java/co/mobileaction/example/web/controller/AdminController.java
@@ -1,18 +1,12 @@
 package co.mobileaction.example.web.controller;
 
-import co.mobileaction.example.web.model.User;
 import co.mobileaction.example.web.service.IPostQueueService;
 import co.mobileaction.example.web.service.IUserQueueService;
 import co.mobileaction.example.web.service.IUserService;
 import co.mobileaction.example.web.util.SecurityUtils;
 import lombok.RequiredArgsConstructor;
-import org.springframework.data.domain.Pageable;
-import org.springframework.data.domain.Sort;
-import org.springframework.data.web.PageableDefault;
-import org.springframework.data.web.SortDefault;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.annotation.Secured;
-import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
@@ -41,16 +35,10 @@ public class AdminController
         return ResponseEntity.ok(true);
     }
 
-    @GetMapping("queue/users")
-    public ResponseEntity<List<User>> getUsers(@PageableDefault(size = 10) @SortDefault(sort = "id", direction = Sort.Direction.ASC) Pageable pageable)
-    {
-        return ResponseEntity.ok(userService.findUsers(pageable));
-    }
-
     @PostMapping("queue/users")
     public ResponseEntity<Boolean> createUserQueueRequests()
     {
-        List<Long> userIds = userService.getDistinctUserIds();
+        List<Long> userIds = userService.findDistinctUserIds();
         userQueueService.sendUserRequests(userIds);
         return ResponseEntity.ok(true);
     }
