+++ web/src/main/java/co/mobileaction/example/web/controller/PostController.java
@@ -42,12 +42,4 @@ public class PostController
 
         return ResponseEntity.ok(true);
     }
-
-    @DeleteMapping("user/{userId}")
-    public ResponseEntity<Boolean> deleteAllPosts(@PathVariable Long userId)
-    {
-        userService.deleteAllPosts(userId);
-
-        return ResponseEntity.ok(true);
-    }
 }
